+++ src/core/design_check.rs
@@ -346,6 +346,7 @@ mod tests {
                 }],
                 ..Default::default()
             },
+            ..Default::default()
         });
         t.devices.push(Device {
             id: "acc1".into(),
@@ -365,6 +366,7 @@ mod tests {
                 }],
                 ..Default::default()
             },
+            ..Default::default()
         });
         t.links.push(Link {
             from: "core".into(),